--- js/___/data/game-data.js
@@ -1,65 +1,0 @@
-export const IMAGES = new Map([
-  [
-    `https://k42.kn3.net/CF42609C8.jpg`,
-    {
-      type: `photo`,
-      width: 600,
-      height: 831
-    }
-  ],
-  [
-    `https://i.imgur.com/DiHM5Zb.jpg`,
-    {
-      type: `photo`,
-      width: 1264,
-      height: 1864
-    }
-  ],
-  [
-    `https://k32.kn3.net/5C7060EC5.jpg`,
-    {
-      type: `photo`,
-      width: 1200,
-      height: 900
-    }
-  ],
-  [
-    `https://k42.kn3.net/D2F0370D6.jpg`,
-    {
-      type: `paint`,
-      width: 468,
-      height: 354
-    }
-  ],
-  [
-    `http://i.imgur.com/1KegWPz.jpg`,
-    {
-      type: `paint`,
-      width: 1080,
-      height: 720
-    }
-  ],
-  [
-    `http://i.imgur.com/DKR1HtB.jpg`,
-    {
-      type: `paint`,
-      width: 1120,
-      height: 2965
-    }
-  ]
-]);
-
-export const ARCHIVE = [
-  {
-    lives: 1,
-    answers: [`wrong`, `slow`, `wrong`, `slow`, `correct`, `slow`, `wrong`]
-  },
-  {
-    lives: 2,
-    answers: [`slow`, `correct`, `correct`, `slow`, `wrong`, `wrong`, `correct`, `wrong`, `correct`, `wrong`]
-  },
-  {
-    lives: 3,
-    answers: [`fast`, `wrong`, `slow`, `wrong`, `slow`, `correct`, `slow`, `wrong`, `fast`, `slow`]
-  }
-];
