+++ js/common/utilites.test.js
@@ -3,7 +3,7 @@ import assert from 'assert';
 // NB! В этом тесте подразумевается, что модуль называется `resize`
 // и экспортирует метод `resize`. Если вы назовёте метод иначе, учтите
 // это в тексте этого теста
-import {resize} from './resize';
+import {resize} from './utilites.js';
 
 const createTestForFrame = (frame) => {
   const assertRatio = (given, expected) => {
