--- js/screen/game/game-view.js
@@ -1,11 +1,0 @@
-import AbstractView from '../../common/abstract-view.js';
-
-export default class GameView extends AbstractView {
-  get template() {
-    return `
-      <section>
-        <p>Gnum a KASHY</p>
-      </section>
-      `;
-  }
-}
