+++ js/screen/game/layout-3-view.js
@@ -1,11 +1,17 @@
-import AbstractView from '../../common/abstract-view.js';
+import Design3View from './design/columns-3-view.js';
+import {IMAGES} from '../../data/data.js';
+import {IMAGE_TYPES} from '../../common/constants.js';
+import {selectImagesPhotoPaint, getTrueAnswer} from '../../common/utilites.js';
 
-export default class Layout3View extends AbstractView {
-  get template() {
-    return `
-      <section>
-        <p>Gnum a KASHY - 3</p>
-      </section>
-      `;
+export default class Layout3View extends Design3View {
+
+  constructor(statistics) {
+
+    super();
+
+    this.title = `Найдите рисунок среди изображений`;
+    this.images = selectImagesPhotoPaint([...IMAGES.keys()], 2, 1);
+    this.trueAnswer = getTrueAnswer(this.images, IMAGE_TYPES.paint);
+    this.statistics = statistics;
   }
 }
