+++ sass/blocks/game.scss
@@ -159,11 +159,11 @@
   content: attr(data-type);
   width: 100px;
   position: absolute;
-  top: 10px;
+  top: -32px;
   left: 50%;
   transform: translate(-50%, 0);
   font-size: 30px;
-  line-height: 38px;
+  line-height: 32px;
   text-align: center;
   color: #de4d51;
   background-color: rgba(120, 120, 120, 0.3);
